tests for codebleu metric v1
